fix test suites their were a completly mess
