--- tests/data/catalog.json
@@ -1,16 +1,0 @@
-{
-    "cid" : "7jz0ygosp3oe",
-    "display_name" : "Catalog    display    name",
-    "items" : [
-        {
-            "value" : "value",
-            "display_name" : "Catalog item 2         name          test",
-            "code" : 2,
-            "description" : "item description",
-            "metadata" : {
-                
-            }
-        }
-    ],
-    "kind" : "SPATIAL"   
-}
